refactor: make loading dots as core:ui component
