+++ core/ui/src/main/java/com/emotionstorage/ui/component/loading/FullLoadingScreen.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ui.component
+package com.emotionstorage.ui.component.loading
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -6,7 +6,9 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.LineHeightStyle
 import androidx.compose.ui.tooling.preview.Preview
 import com.emotionstorage.ui.theme.MooiTheme
 
@@ -24,8 +26,9 @@ fun FullLoadingScreen(modifier: Modifier = Modifier) {
                     .fillMaxSize()
                     .background(MooiTheme.colorScheme.background)
                     .padding(innerPadding),
+            contentAlignment = Alignment.Center,
         ) {
-            // todo: add loading ui
+            LoadingDots()
         }
     }
 }
