+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/ui/component/ChatMessageList.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.unit.dp
 import com.emotionstorage.domain.model.ChatMessage
 import com.emotionstorage.domain.model.ChatMessage.MessageSource
 import com.emotionstorage.common.toKorDateWithWeekDay
+import com.emotionstorage.ui.component.loading.LoadingDots
 import com.emotionstorage.ui.theme.MooiTheme
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -105,7 +106,7 @@ fun ChatMessageList(
                                 ),
                         contentAlignment = Alignment.CenterStart,
                     ) {
-                        MessageIncomeLoading()
+                        LoadingDots()
                     }
                 }
             }
